requirements.txt
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
conllu==4.2
decorator==4.4.2
dgl-cu102==0.5.3
filelock==3.0.12
fire==0.3.1
future==0.18.2
h5py==2.10.0
idna==2.10
joblib==0.16.0
jsonlines==1.2.0
Keras==2.4.3
networkx==2.5
nltk==3.5
numpy==1.19.2
packaging==20.4
pyparsing==2.4.7
PyYAML==5.3.1
regex==2020.7.14
requests==2.24.0
sacremoses==0.0.43
scipy==1.5.2
sentencepiece==0.1.91
seqeval==0.0.12
six==1.15.0
termcolor==1.1.0
tokenizers==0.8.1rc1
torch==1.6.0
tqdm==4.49.0
transformers==3.0.2
urllib3==1.25.10
